Move logic out of console
