+++ ert3/__init__.py
@@ -1,6 +1,8 @@
 import ert3.console
 import ert3.storage
 import ert3.evaluator
+import ert3.engine
+import ert3.workspace
 
 
 _WORKSPACE_DATA_ROOT = ".ert"
