+++ ert3/workspace/__init__.py
@@ -1,0 +1,4 @@
+from ert3.workspace._workspace import initialize
+from ert3.workspace._workspace import load 
+from ert3.workspace._workspace import experiment_have_run
+from ert3.workspace._workspace import assert_experiment_exists
